swell-gtk-im: move swell-im code into swell-wnd-generic.cpp to remove duplicate code/structure definitions
